Создает PointController и перемещает в него смену точки маршрута на форму редактирования
